callsAPI/requirements.txt
Django==2.1
djangorestframework==3.8.2
psycopg2==2.7.5
dj_database_url==0.5.0
gunicorn==19.9.0
django-rest-swagger==2.2.0
django-filter==2.0.0